refactor(recog): share ensure_binary_with_threshold helper
classapp.rs と pageseg.rs にほぼ同じ ensure_binary_with_threshold が
重複していたため、src/recog/util.rs に集約した。pageseg 側は 32bpp 入力
に対して独自の整数 luminance (0.301, 0.586, 0.113) を使っていたが、
classapp 側のように Pix::convert_to_8 経由 (0.3, 0.5, 0.2) に統一する。
任意 depth (2/4/16bpp 等) も convert_to_8 で正しく扱えるようになる。
